wallet modal - override styles
