+++ src/components/organisms/Modals/index.tsx
@@ -1,10 +1,12 @@
 import React, { Fragment } from "react";
 import ModalConnectWallet from "./ModalConnectWallet/modalConnectWallet.organism";
+import ModalConnectWalletV2 from "./ModalConnectWalletV2/modalConnectWalletV2.organism";
 
 const Modals: React.FC = () => {
   return (
     <Fragment>
       <ModalConnectWallet />
+      <ModalConnectWalletV2 />
     </Fragment>
   );
 };
